Moved TargetEndpointInfo and TargetVideoPlayerInfo out of tv-casting-app/linux/main.cpp Refactored tv-casting-app/linux/main.cpp: Created CastingServer, renamed Casting->CastingUtils
