+++ examples/tv-casting-app/linux/BUILD.gn
@@ -23,9 +23,10 @@ assert(chip_build_tools)
 executable("chip-tv-casting-app") {
   sources = [
     "${chip_root}/examples/tv-casting-app/tv-casting-common/include/CHIPProjectAppConfig.h",
-    "Casting.h",
     "CastingShellCommands.cpp",
     "CastingShellCommands.h",
+    "CastingUtils.cpp",
+    "CastingUtils.h",
     "main.cpp",
   ]
 
@@ -36,6 +37,9 @@ executable("chip-tv-casting-app") {
     "${chip_root}/src/lib",
     "${chip_root}/third_party/inipp",
   ]
+
+  include_dirs =
+      [ "${chip_root}/examples/tv-casting-app/tv-casting-common/include" ]
 
   cflags = [ "-Wconversion" ]
